requirements.txt
broadlink>=0.18.0
requests>=2.28.0
PyYAML>=6.0
